src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
supabase==2.9.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
openai>=1.40.0
celery[redis]>=5.4.0
